homecheck/requirements_local.txt
boto3
dj-database-url
Django
django-celery-beat
django-filter
django-redis
django-storages
djangorestframework
gunicorn
Markdown
numpy
pillow
redis
pyTelegramBotAPI
python-dotenv
openai
drf-yasg
django-cors-headers